Write read Mongo test
